pyproject.toml
[project]
name = "diaspy"
version = "0.1.0"
description = "A DSPy-based package for multi-LLM dialectical workflows, solving the AI management problem through philosophically inspired agent interactions."
authors = [{ name = "The SciPhi Initiative, LLC" }]
readme = "README.md"
requires-python = ">=3.8"
dependencies = ["dspy-ai"]
license = { text = "Copyright (c) 2023 The SciPhi Initiative, LLC. All rights reserved." }

[project.scripts]
diaspy = "diaspy.cli:main"

[project.optional-dependencies]
dev = ["pytest", "black", "ruff"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.poetry]  # If using Poetry; can be removed if preferring another build system
name = "diaspy"
version = "0.1.0"
description = "A DSPy-based package for multi-LLM dialectical workflows, solving the AI management problem through philosophically inspired agent interactions."
authors = ["The SciPhi Initiative, LLC"]
readme = "README.md"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
